+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Change Log
 ==========
 
+Version 0.6.0 *(3rd September 2021)*
+-------------------------------------------
+- Adds public methods for suspending/discarding & resuming InApp Notifications
+- Adds public methods to increment/decrement values set via User properties
+- Deprecates `profileGetCleverTapID()` and `profileGetCleverTapAttributionIdentifier()`
+- Adds a new public method `getCleverTapID()` as an alternative to above deprecated methods
+- Supports CleverTap iOS SDK `v3.10.0`
+
+Version 0.5.2 *(20th July 2021)*
+-------------------------------------------
+- Supports CleverTap Android SDK `v4.2.0`
+- Android SDK `v4.2.0` fixes NPE for `recordScreenView()` in Android
+
 Version 0.5.1 *(5th May 2021)*
 -------------------------------------------
 - Update and Freeze [CleverTap React Native Podspec](/clevertap-react-native.podspec) to a specific version of a CleverTap iOS SDK